requirements.txt
# ══════════════════════════════════════════════════════════
# ScholarMind — Requirements
# ══════════════════════════════════════════════════════════
# Install: pip install -r requirements.txt

# ── Core agents ───────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.33.0
litellm>=1.40.0

# ── UI ────────────────────────────────────────────────────
streamlit>=1.40.0

# ── LLM providers ─────────────────────────────────────────
openai>=1.50.0
groq>=0.9.0

# ── LangChain ─────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0

# ── Vector store — FAISS (same as original notebook) ──────
faiss-cpu>=1.8.0

# ── Embeddings — free, no API key needed ──────────────────
sentence-transformers>=3.0.0

# ── Web crawling ──────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0

# ── Data ─────────────────────────────────────────────────
pandas>=2.0.0
python-dotenv>=1.0.0

# ── Optional: HuggingFace LLM (if no Groq/OpenAI key) ────
# transformers>=4.40.0
# torch>=2.0.0
# accelerate>=0.27.0
